Creating class structure and adding DanmEp sorting code.
The imagined flow is first sorting applicable network policies and DanmEps,
then parsing an pplicable set of rules from the two.
Finally the rules which need to be applied are passed to a provisioner via an abstract interface.
The first such provisioner will be obviously "iptables", but a future, better performing eBPF based
based provisioner is also very much in the picture.