Cleaner: reconciling the handling of "no release service" and "no IP" cases
DANM IPAM assumes "empty" IPs are not "allocated" by it, which is technically true.
Cleaner treated this neutral use-case as an error, which actually resulted in un-related addresses remaining non-cleaned,
and DanmEps being unnecessarily kept.
Handling is changed to a more future-proof approach: we don't treat non-registered release implemnetations as errors, and failing to clean one IP will not block the release process of another anymore.